+++ test/datetime/tokenParse.test.js
@@ -3,7 +3,8 @@ import { DateTime } from "../../src/luxon";
 import {
   ConflictingSpecificationError,
   UnparsableStringError,
-  MismatchedWeekdayError
+  MismatchedWeekdayError,
+  UnitOutOfRangeError
 } from "../../src/errors";
 
 //------
@@ -222,6 +223,21 @@ test("DateTime.fromFormat() parses format month names", () => {
   expect(i.day).toBe(25);
 });
 
+test("DateTime.fromFormat() parses quarters", () => {
+  const i = DateTime.fromFormat("1982Q2", "yyyy[Q]q");
+  expect(i.year).toBe(1982);
+  expect(i.month).toBe(4);
+  expect(i.quarter).toBe(2);
+  expect(DateTime.fromFormat("2019Q1", "yyyy[Q]q").month).toBe(1);
+  expect(DateTime.fromFormat("2019Q2", "yyyy[Q]q").month).toBe(4);
+  expect(DateTime.fromFormat("2019Q3", "yyyy[Q]q").month).toBe(7);
+  expect(DateTime.fromFormat("2019Q4", "yyyy[Q]q").month).toBe(10);
+  expect(DateTime.fromFormat("2019Q01", "yyyy[Q]qq").month).toBe(1);
+  expect(DateTime.fromFormat("2019Q02", "yyyy[Q]qq").month).toBe(4);
+  expect(DateTime.fromFormat("2019Q03", "yyyy[Q]qq").month).toBe(7);
+  expect(DateTime.fromFormat("2019Q04", "yyyy[Q]qq").month).toBe(10);
+});
+
 test("DateTime.fromFormat() makes trailing periods in month names optional", () => {
   const i = DateTime.fromFormat("janv 25 1982", "LLL dd yyyy", {
     locale: "fr"
@@ -381,10 +397,16 @@ test("DateTime.fromFormat() rejects gibberish", () => {
   expect(() => DateTime.fromFormat("Splurk", "EEEE")).toThrow(UnparsableStringError);
 });
 
+test("DateTime.fromFormat() rejects invalid quarter value", () => {
+  expect(() => DateTime.fromFormat("2019Qaa", "yyyy[Q]qq")).toThrow(UnparsableStringError);
+  expect(() => DateTime.fromFormat("2019Q00", "yyyy[Q]qq")).toThrow(UnitOutOfRangeError);
+  expect(() => DateTime.fromFormat("2019Q0", "yyyy[Q]q")).toThrow(UnitOutOfRangeError);
+  expect(() => DateTime.fromFormat("2019Q5", "yyyy[Q]q")).toThrow(UnitOutOfRangeError);
+});
+
 test("DateTime.fromFormat() rejects out-of-range values", () => {
   // todo - these are actually several different kinds of errors. clean this up
   const rejects = (s, fmt, opts = {}) => expect(() => DateTime.fromFormat(s, fmt, opts)).toThrow();
-
   rejects("8, 05/25/1982", "E, MM/dd/yyyy", { locale: "fr" });
   rejects("Tuesday, 05/25/1982", "EEEE, MM/dd/yyyy", { locale: "fr" });
   rejects("Giberish, 05/25/1982", "EEEE, MM/dd/yyyy");