+++ test/interval/format.test.js
@@ -23,6 +23,25 @@ test("Interval#toISO accepts ISO options", () =>
   expect(interval.toISO({ suppressSeconds: true })).toBe("1982-05-25T09:00Z/1983-10-14T13:30Z"));
 
 //------
+// .toISODate()
+//------
+
+test("Interval#toISODate returns a simple ISO date interval format", () =>
+  expect(interval.toISODate()).toBe("1982-05-25/1983-10-14"));
+
+//------
+// .toISOTime()
+//------
+
+test("Interval#toISOTime returns a simple ISO time interval format", () =>
+  expect(interval.toISOTime()).toBe("09:00:00.000Z/13:30:00.000Z"));
+
+test("Interval#toISOTime accepts ISO options", () => {
+  expect(interval.toISOTime({ suppressSeconds: true })).toBe("09:00Z/13:30Z");
+  expect(interval.toISOTime({ suppressMilliseconds: true })).toBe("09:00:00Z/13:30:00Z");
+});
+
+//------
 // .toFormat()
 //------